Извлек Recyclers Adapter в отдельные классы
