+++ app/src/main/java/com/mk/rickmortyappbykrautsevich/fragments/CharactersListFragment.kt
@@ -2,7 +2,6 @@ package com.mk.rickmortyappbykrautsevich.fragments
 
 import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -10,15 +9,16 @@ import android.view.ViewGroup
 import android.widget.*
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.mk.rickmortyappbykrautsevich.FragmentHost
 import com.mk.rickmortyappbykrautsevich.HasBottomNavs
 import com.mk.rickmortyappbykrautsevich.R
 import com.mk.rickmortyappbykrautsevich.enums.Gender
 import com.mk.rickmortyappbykrautsevich.enums.Status
-import com.mk.rickmortyappbykrautsevich.fragments.recyclers_data.CharacterRecData
+import com.mk.rickmortyappbykrautsevich.fragments.recyclers.CharacterAdapter
+import com.mk.rickmortyappbykrautsevich.fragments.recyclers_data.CharacterData
 import com.mk.rickmortyappbykrautsevich.retrofit.models.queries.CharacterQuery
 import com.mk.rickmortyappbykrautsevich.viewmodels.AllCharactersViewModel
 import com.squareup.picasso.Picasso
@@ -31,6 +31,7 @@ class CharactersListFragment : Fragment() {
     }
 
     private var hasBottomNavs: HasBottomNavs? = null
+    private var host: FragmentHost? = null
     private val picasso = Picasso.get()
     private var recyclerView: RecyclerView? = null
     private var noResults: TextView? = null
@@ -52,12 +53,14 @@ class CharactersListFragment : Fragment() {
 
     private var loadingLiveData: LiveData<Boolean>? = null
     private var pagingLiveData: LiveData<Boolean>? = null
-    private var charactersLiveData: LiveData<List<CharacterRecData>>? = null
+    private var charactersLiveData: LiveData<List<CharacterData>>? = null
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
         if (context is HasBottomNavs)
             hasBottomNavs = context
+        if (context is FragmentHost)
+            host = context
     }
 
     override fun onCreateView(
@@ -124,14 +127,16 @@ class CharactersListFragment : Fragment() {
                 }
             })
 
-            adapter = CharacterAdapter(ArrayList())
+            adapter = CharacterAdapter(ArrayList(), picasso, host)
         }
         charactersLiveData?.observe(
             viewLifecycleOwner
         ) { list ->
             list?.let {
                 (recyclerView?.adapter as CharacterAdapter).changeContacts(it)
-                if (it.isNotEmpty()) {showRecycler()} else showNoResults()
+                if (it.isNotEmpty()) {
+                    showRecycler()
+                } else showNoResults()
             }
         }
         initShowButtonListener()
@@ -141,6 +146,7 @@ class CharactersListFragment : Fragment() {
     override fun onDetach() {
         super.onDetach()
         hasBottomNavs = null
+        host = null
     }
 
     private fun initSpinners() {
@@ -227,19 +233,17 @@ class CharactersListFragment : Fragment() {
                     gender = gender?.name,
                     status = status?.name
                 )
-                Log.d("12345", query.gender.toString())
-                Log.d("12345", query.status.toString())
             }
             viewModel.getData(query)
         }
     }
 
-    private fun showRecycler(){
+    private fun showRecycler() {
         recyclerView?.visibility = View.VISIBLE
         noResults?.visibility = View.INVISIBLE
     }
 
-    private fun showNoResults(){
+    private fun showNoResults() {
         recyclerView?.visibility = View.INVISIBLE
         noResults?.visibility = View.VISIBLE
     }
@@ -249,79 +253,6 @@ class CharactersListFragment : Fragment() {
         swipe.setOnRefreshListener {
             viewModel.getData(null)
             swipe.isRefreshing = false
-        }
-    }
-
-    inner class CharacterAdapter(private var characters: List<CharacterRecData>) :
-        RecyclerView.Adapter<CharacterAdapter.CharacterHolder>() {
-
-        override fun onCreateViewHolder(
-            parent: ViewGroup,
-            viewType: Int
-        ): CharacterAdapter.CharacterHolder {
-            val itemView = LayoutInflater.from(parent.context)
-                .inflate(R.layout.holder_character, parent, false)
-            return CharacterHolder(itemView)
-        }
-
-        override fun onBindViewHolder(holder: CharacterAdapter.CharacterHolder, position: Int) {
-            holder.bind(characters[position])
-        }
-
-        override fun getItemCount(): Int = characters.size
-
-        fun changeContacts(list: List<CharacterRecData>) {
-            val new = ArrayList<CharacterRecData>()
-            new.addAll(list)
-            val old = characters
-            val diffUtilCallback = ContactDiffUtilCallBack(oldList = old, newList = new)
-            val result = DiffUtil.calculateDiff(diffUtilCallback, false)
-            characters = new
-            result.dispatchUpdatesTo(this)
-        }
-
-        inner class CharacterHolder(itemView: View) : RecyclerView.ViewHolder(itemView),
-            View.OnClickListener {
-            private val image: ImageView = itemView.findViewById(R.id.char_image)
-            private val nameTextView: TextView = itemView.findViewById(R.id.name_textview)
-            private val specTextView: TextView = itemView.findViewById(R.id.spec_textview)
-            private val genderTextView: TextView = itemView.findViewById(R.id.gender_textview)
-            private val statusTextView: TextView = itemView.findViewById(R.id.status_textview)
-            private var characterBound: CharacterRecData? = null
-
-            init {
-                itemView.setOnClickListener { this }
-            }
-
-            fun bind(character: CharacterRecData) {
-                characterBound = character
-                nameTextView.text = character.name
-                specTextView.text = character.species
-                genderTextView.text = character.gender
-                statusTextView.text = character.status
-                picasso.load(character.image).into(image)
-            }
-
-            override fun onClick(p0: View?) {
-                TODO("Not yet implemented")
-            }
-
-        }
-
-        inner class ContactDiffUtilCallBack(
-            private val oldList: List<CharacterRecData>,
-            private val newList: List<CharacterRecData>
-        ) : DiffUtil.Callback() {
-            override fun getOldListSize() = oldList.size
-
-            override fun getNewListSize() = newList.size
-
-            override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int) =
-                (oldList[oldItemPosition].id == newList[newItemPosition].id)
-
-            override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-                return oldList[oldItemPosition] == newList[newItemPosition]
-            }
         }
     }
 }
