+++ app/src/main/java/com/mk/rickmortyappbykrautsevich/fragments/EpisodeListFragment.kt
@@ -13,13 +13,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.mk.rickmortyappbykrautsevich.FragmentHost
 import com.mk.rickmortyappbykrautsevich.HasBottomNavs
 import com.mk.rickmortyappbykrautsevich.R
+import com.mk.rickmortyappbykrautsevich.fragments.recyclers.EpisodeAdapter
 import com.mk.rickmortyappbykrautsevich.fragments.recyclers_data.EpisodeData
 import com.mk.rickmortyappbykrautsevich.retrofit.models.queries.EpisodeQuery
 import com.mk.rickmortyappbykrautsevich.viewmodels.AllEpisodesViewModel
@@ -32,7 +32,7 @@ class EpisodeListFragment : Fragment() {
     }
 
     private var hasBottomNavs: HasBottomNavs? = null
-    private var fragmentHost: FragmentHost? = null
+    private var host: FragmentHost? = null
 
     private var recyclerView: RecyclerView? = null
     private var noResults: TextView? = null
@@ -56,7 +56,7 @@ class EpisodeListFragment : Fragment() {
         if (context is HasBottomNavs)
             hasBottomNavs = context
         if (context is FragmentHost)
-            fragmentHost = context
+            host = context
     }
 
     override fun onCreateView(
@@ -124,11 +124,11 @@ class EpisodeListFragment : Fragment() {
                 }
             })
 
-            adapter = EpisodeAdapter(ArrayList())
+            adapter = EpisodeAdapter(ArrayList(), host)
         }
         episodeLiveData?.observe(viewLifecycleOwner) { list ->
             list?.let {
-                (recyclerView?.adapter as EpisodeListFragment.EpisodeAdapter).changeContacts(it)
+                (recyclerView?.adapter as EpisodeAdapter).changeContacts(it)
                 if (it.isNotEmpty()) {
                     showRecycler()
                 } else showNoResults()
@@ -141,6 +141,7 @@ class EpisodeListFragment : Fragment() {
     override fun onDetach() {
         super.onDetach()
         hasBottomNavs = null
+        host = null
     }
 
     private fun initShowButtonListener() {
@@ -188,75 +189,6 @@ class EpisodeListFragment : Fragment() {
         swipe.setOnRefreshListener {
             viewModel.getData(null)
             swipe.isRefreshing = false
-        }
-    }
-
-    inner class EpisodeAdapter(private var episodes: List<EpisodeData>) :
-        RecyclerView.Adapter<EpisodeAdapter.EpisodeHolder>() {
-
-        override fun onCreateViewHolder(
-            parent: ViewGroup,
-            viewType: Int
-        ): EpisodeAdapter.EpisodeHolder {
-            val itemView = LayoutInflater.from(parent.context)
-                .inflate(R.layout.holder_episode, parent, false)
-            return EpisodeHolder(itemView)
-        }
-
-        override fun onBindViewHolder(holder: EpisodeAdapter.EpisodeHolder, position: Int) {
-            holder.bind(episodes[position])
-        }
-
-        override fun getItemCount(): Int = episodes.size
-
-        fun changeContacts(list: List<EpisodeData>) {
-            val new = ArrayList<EpisodeData>()
-            new.addAll(list)
-            val old = episodes
-            val diffUtilCallback = ContactDiffUtilCallBack(oldList = old, newList = new)
-            val result = DiffUtil.calculateDiff(diffUtilCallback, false)
-            episodes = new
-            result.dispatchUpdatesTo(this)
-        }
-
-        inner class EpisodeHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-            private val nameTextView: TextView = itemView.findViewById(R.id.ep_name_textview)
-            private val numberTextView: TextView = itemView.findViewById(R.id.ep_number_textview)
-            private val dateTextView: TextView = itemView.findViewById(R.id.ep_air_date_textview)
-            private var episodeBound: EpisodeData? = null
-
-            init {
-                itemView.setOnClickListener {
-                    val fragment =
-                        episodeBound?.id?.let { EpisodeDetailFragment.newInstance(it) }
-                    fragment?.let {
-                        fragmentHost?.setFragment(it)
-                    }
-                }
-            }
-
-            fun bind(episode: EpisodeData) {
-                episodeBound = episode
-                nameTextView.text = episode.name
-                numberTextView.text = episode.episode
-                dateTextView.text = episode.airDate
-            }
-        }
-
-        inner class ContactDiffUtilCallBack(
-            private val oldList: List<EpisodeData>,
-            private val newList: List<EpisodeData>
-        ) : DiffUtil.Callback() {
-            override fun getOldListSize() = oldList.size
-
-            override fun getNewListSize() = newList.size
-
-            override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int) =
-                (oldList[oldItemPosition].id == newList[newItemPosition].id)
-
-            override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-                return oldList[oldItemPosition] == newList[newItemPosition]
-            }
         }
     }
 }
