+++ app/src/main/res/layout/fragment_episode_detail.xml
@@ -137,6 +137,16 @@
             app:layout_constraintBottom_toBottomOf="parent"
             android:visibility="visible"/>
 
+        <ProgressBar
+            android:id="@+id/list_progress_bar"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="@id/recycler"
+            app:layout_constraintEnd_toEndOf="@id/recycler"
+            app:layout_constraintTop_toTopOf="@id/recycler"
+            app:layout_constraintBottom_toBottomOf="@id/recycler"
+            android:visibility="invisible"/>
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
